Brick Spawning Level complete
When I changed the -row_width it didn't change the spawning issue, the last row and the last column are always too big, sometimes I get bigger bricks too after a few collisions, I rechecked, but not sure what the error is. I think I'll make a branch from here and continue to see if the other, more complex brick spawner works in the end.